requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
idna==3.10
magic-filter==1.0.12
multidict==6.3.2
passlib==1.7.4
pillow==11.2.0
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0
yarl==1.19.0
